TEAL-539: Add lint, test, build and documentation stages to CI pipeline
Note that there is an (undesireable) divergence between the base images used for developing and testing. The development image uses Node 21.7.1 and the test image uses Node 20.12.2. This is the most recent node version available in the Cypress browser test images. Ideally, this gap should be closed in the future.