DAAC-140: Single pulse candidate labelling
Added a SPLabeller component which supports assignment of RFI, single pulse and pulsar labels to groups of candidates. Once labels are assigned, they are plotted in a different color. Each group of entities can be toggled by clicking the corresponding entity on the plot key. Re-labelling a candidate overwrites the previous label.