DAAC-93: Update Vite build config
* Configures Rollup manual chunks in Vite config
* Ignores 'Can't resolve original location of error.' log messages during vite build

See https://github.com/vitejs/vite/issues/15012#issuecomment-1815854072 for a discussion about the issue and the workaround. The issue is present when Rollup (module bundler for javascript, used by vite build) attempts to build tanstack-query and MUI dependencies with the vite config option build.sourcemap=true. Build.sourcemap needs to be true in order for code coverage to work as expected.

Also see https://rollupjs.org/configuration-options/#output-manualchunks. Allows creation of custom shared common chunks. The functional form is used, which means that each resolved module id will be passed to the function and the module name is returned as a string. The module and all its dependency will be added to the manual chunk with the given name.

Without the workaround, vite build is oom-killed. Another possible workaround is to increase the maximum node memory usage (e.g. export NODE_OPTIONS=--max-old-space-size=<some large number of MB>). This is undesireable because the gitlab pipeline runs in K8s containers, which likely don't have significantly more more to allocate.

It's therefore recommended to set export NODE_OPTIONS=--max-old-space-size=N where N <= 2048 in development to match the memory limit of SKAO Gitlab runners.