DAAC-318: Refactor auth and login redirect functions
The root route (/) is configured to redirect to SKA-IAM for user login. If the user has a valid session (i.e. they have already authenticated with SKA-IAM) then they will be redirected to the /label route.

If the VITE_AUTH_DISABLED environment variable is set to 'true' then the redirect from / to SKA-IAM is bypassed. This effectively disables the authentication flow in the application.

To support these changes, login and authentication functions have been moved out of the LoginButton component and into their own modules.