DAAC-318: Force logout on auth error response
The idea here is that the frontend relies on the backend throwing an error to understand if the user's session has been revoked. A user's access token is sent in request headers to be validated in the backend. If the token has expired or is invalid, the backend returns a HTTP 401 (unauthorized) response. Note that HTTP 401 tells us the user is not authenticated. This is different to a HTTP 403 (forbidden) response which tells us that a user isn't allowed to access the requested resource.

An axios response interceptor is configured to check if the backend returns a HTTP 401 response. If it does, then the user's access token is cleared from local storage. This means that the access token isn't embedded into subsequent requests by the axios request interceptor, which effictively unauthenticates the user. Note that this approach doesn't actually terminate the session or invalidate the token.